# frontend/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quant Platform - Quant-Service [Development]",
    "description": "Data Service API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/market-data/intelligence/news/{symbol}": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "intelligence"
        ],
        "summary": "종목 뉴스 조회",
        "description": "지정된 종목 관련 뉴스를 조회합니다.",
        "operationId": "Market Data v2-get_news",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "종목 심볼 (예: AAPL, TSLA)",
              "title": "Symbol"
            },
            "description": "종목 심볼 (예: AAPL, TSLA)"
          },
          {
            "name": "topics",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "관심 주제 (쉼표로 구분)",
              "title": "Topics"
            },
            "description": "관심 주제 (쉼표로 구분)"
          },
          {
            "name": "time_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "시작 날짜 (YYYY-MM-DD)",
              "title": "Time From"
            },
            "description": "시작 날짜 (YYYY-MM-DD)"
          },
          {
            "name": "time_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "종료 날짜 (YYYY-MM-DD)",
              "title": "Time To"
            },
            "description": "종료 날짜 (YYYY-MM-DD)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "결과 개수 제한",
              "default": 50,
              "title": "Limit"
            },
            "description": "결과 개수 제한"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get News"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/intelligence/sentiment/{symbol}": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "intelligence"
        ],
        "summary": "감정 분석 조회",
        "description": "지정된 종목의 감정 분석 결과를 조회합니다.",
        "operationId": "Market Data v2-get_sentiment_analysis",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "종목 심볼 (예: AAPL, TSLA)",
              "title": "Symbol"
            },
            "description": "종목 심볼 (예: AAPL, TSLA)"
          },
          {
            "name": "time_range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "시간 범위 (1week, 1month, 3months, 6months)",
              "default": "1month",
              "title": "Time Range"
            },
            "description": "시간 범위 (1week, 1month, 3months, 6months)"
          },
          {
            "name": "sources",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "분석 소스 (쉼표로 구분)",
              "title": "Sources"
            },
            "description": "분석 소스 (쉼표로 구분)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Sentiment Analysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/intelligence/analyst-recommendations/{symbol}": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "intelligence"
        ],
        "summary": "분석가 추천 조회",
        "description": "지정된 종목의 분석가 추천 정보를 조회합니다.",
        "operationId": "Market Data v2-get_analyst_recommendations",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "종목 심볼 (예: AAPL, TSLA)",
              "title": "Symbol"
            },
            "description": "종목 심볼 (예: AAPL, TSLA)"
          },
          {
            "name": "time_range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "시간 범위 (1month, 3months, 6months, 1year)",
              "default": "3months",
              "title": "Time Range"
            },
            "description": "시간 범위 (1month, 3months, 6months, 1year)"
          },
          {
            "name": "brokers",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "증권사 리스트 (쉼표로 구분)",
              "title": "Brokers"
            },
            "description": "증권사 리스트 (쉼표로 구분)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Analyst Recommendations"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/intelligence/social-sentiment/{symbol}": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "intelligence"
        ],
        "summary": "소셜 미디어 감정 분석 조회",
        "description": "지정된 종목의 소셜 미디어 감정 분석을 조회합니다.",
        "operationId": "Market Data v2-get_social_sentiment",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "종목 심볼 (예: AAPL, TSLA)",
              "title": "Symbol"
            },
            "description": "종목 심볼 (예: AAPL, TSLA)"
          },
          {
            "name": "platforms",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "소셜 미디어 플랫폼 (twitter,reddit,stocktwits)",
              "title": "Platforms"
            },
            "description": "소셜 미디어 플랫폼 (twitter,reddit,stocktwits)"
          },
          {
            "name": "time_range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "시간 범위 (1day, 1week, 1month)",
              "default": "1week",
              "title": "Time Range"
            },
            "description": "시간 범위 (1day, 1week, 1month)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Social Sentiment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}